requirements.txt
python-telegram-bot
Pillow
nest_asyncio
